src/azureml-studio-model/requirements.txt
# https://pip.pypa.io/en/stable/reference/pip_install/#requirements-file-format
pandas
cloudpickle
PyYAML
pytest

# WARNING: do NOT upgrade this version because azureml-core 1.0.48 requires it to be <=0.15.89,>=0.15.35
ruamel.yaml==0.15.89